Add readLineFromStdin and remove hi.c and hi.h
Add readLineFromStdin so we can create a lttng session and get a listing
of the tracepoints and know that things are ok. Its not necessary to have
the readLineFromStdin you can enable-events before running an application
but this is the method described for a instrumenting a hello world
program with lttng.